+++ lib/src/compiler/pipe_resolver.dart
@@ -1,8 +1,7 @@
 import "package:angular2/src/core/di.dart" show Injectable;
 import "package:angular2/src/core/metadata.dart" show Pipe;
-import "package:angular2/src/core/reflection/reflection.dart" show reflector;
-import "package:angular2/src/core/reflection/reflector_reader.dart"
-    show ReflectorReader;
+import "package:angular2/src/core/reflection/reflection.dart"
+    show Reflector, reflector;
 import "package:angular2/src/facade/exceptions.dart" show BaseException;
 
 /// Resolve a type for [Pipe].
@@ -13,8 +12,8 @@ import "package:angular2/src/facade/exceptions.dart" show BaseException;
 /// See [Compiler]
 @Injectable()
 class PipeResolver {
-  ReflectorReader _reflector;
-  PipeResolver([ReflectorReader _reflector]) {
+  Reflector _reflector;
+  PipeResolver([Reflector _reflector]) {
     this._reflector = _reflector ?? reflector;
   }