+++ lib/src/core/reflection/reflection.dart
@@ -6,18 +6,29 @@ import 'types.dart';
 
 export 'reflector.dart';
 
+/// An implementation of reflection capabilities that throws when used.
+///
+/// By default, [reflector] is setup to use [NoReflectionCapabilities]. To
+/// opt-in to runtime reflection:
+///     import 'package:angular2/reflection.dart';
+///
+///     void main() {
+///       allowRuntimeReflection();
+///     }
 class NoReflectionCapabilities implements PlatformReflectionCapabilities {
+  const NoReflectionCapabilities._();
+
   @override
-  bool isReflectionEnabled() {
-    return false;
-  }
+  bool isReflectionEnabled() => true;
 
   @override
   Function factory(Type type) =>
       throw new NoReflectionCapabilitiesError._noInfo(type);
+
   @override
   List interfaces(Type type) =>
       throw new NoReflectionCapabilitiesError._noInfo(type);
+
   @override
   List<List> parameters(dynamic type) =>
       throw new NoReflectionCapabilitiesError._noInfo(type);
@@ -29,6 +40,7 @@ class NoReflectionCapabilities implements PlatformReflectionCapabilities {
   @override
   Map<String, List> propMetadata(dynamic type) =>
       throw new NoReflectionCapabilitiesError._noInfo(type);
+
   @override
   GetterFn getter(String name) =>
       throw new NoReflectionCapabilitiesError._("Cannot find getter ${name}");
@@ -45,8 +57,12 @@ class NoReflectionCapabilities implements PlatformReflectionCapabilities {
   String importUri(Type type) => './';
 }
 
-final Reflector reflector = new Reflector(new NoReflectionCapabilities());
+/// Global reflector shared by all Angular applications.
+///
+/// This is considered to be **internal only**.
+final Reflector reflector = new Reflector(const NoReflectionCapabilities._());
 
+/// Exception thrown by [NoReflectionCapabilities].
 class NoReflectionCapabilitiesError extends Error {
   final String message;